Add comparison (#6)
Continued fractions are orderable, so should ContinuedFractions.

This commit mixes in the Comparable module and defines the spaceship
operator, so ContinuedFractions can be ordered.

Comparison is restricted to `ContinuedFraction#number`.  `#limit`
shouldn't be part of the comparison, since precision isn't lost in
operations between ContinuedFractions, i.e. the larger of the two limits
is is returned in the result.