Fix floating point representation error bug (#7)
The quotients were not being calculated accurately for certain rational
numbers. For example,

ContinuedFraction.new(7/4r)

produced the quotients..

[1, 1, 3, 1125899906842624]

and convergents..

[[1, 1], [2, 1], [7, 4], [7881299347898370, 4503599627370497]]

The issue is caused by floating point representation errors, which seem to
impact most languages, not just Ruby. A simpler example is..

1.4 + 1.2
=> 2.5999999999999996

These representation issues impact the continued fraction quotients.

This change addresses the bug, by adding a second break condition in the
quotients algorithm. The condition looks to see if the currently generated
convergent is equal to the number being converted.

In order to break when this condition is met, the code in calculate_quotients
and calculate_convergents was merged into one method. The merge allows to
inline the break conditions as the quotients are being calculated and avoid
extra clean-up code.

Some info can be found here..
https://corainchicago.github.io/blog/why-does-ruby-add-numbers-wrong/