Streamline code and fix a computation bug
We can make the API more concise and the supporting logic in CF simpler, if
we store Rationals in the convergents array, instead of Array tuples.

The method #convergents_as_rationals, can be eliminated if we store Rationals
in the convergents array, in place of Array tuples. A call to #convergents
returns an array of Rationals. This is more concise and simpler logic to
support.

This commit replaces array tuples for Rationals in the convergents array.

It also addresses a bug where convergents for finite decimal numbers and
fractions would be calculated erroneously beyond the required range. For
example, the convergents array for (7/6r).to_f would be:

[[1, 1], [6, 5], [7, 6], [2627099782632793, 2251799813685251], [7881299347898386, 6755399441055759]]

This fix is:
  1) to do all calculations using Rational type
  2) to break the loop when an epsilon of zero between the number and the
     current convergent is encountered

Lastly, the documentation is now being generated by Yard. Headers were updated
to support this. Github pages and RubyGems links will be updated to point to
the new documentation.